% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{Rtskit-package}
\alias{Rtskit}
\alias{Rtskit-package}
\title{Rtskit: native access to succinct tree sequences}
\description{
Rtskit directly wraps the bundled, pinned tskit C API. It has no Python,
Rust, Java, or external shared-library runtime dependency.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/RGenomicsETL/Rtskit}
  \item Report bugs at \url{https://github.com/RGenomicsETL/Rtskit/issues}
}

}
\author{
\strong{Maintainer}: Sounkou Mahamane Toure \email{sounkoutoure@gmail.com}

Authors:
\itemize{
  \item Sounkou Mahamane Toure \email{sounkoutoure@gmail.com}
}

Other contributors:
\itemize{
  \item Tskit Developers (Bundled tskit C library) [copyright holder]
}

}
\keyword{internal}
