% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-sequence.R
\name{tskit_edges}
\alias{tskit_edges}
\title{Materialize the edge table}
\usage{
tskit_edges(x)
}
\arguments{
\item{x}{A \code{TreeSequence}.}
}
\value{
A data frame with zero-based IDs and edge coordinates and nodes.
}
\description{
Metadata remains native and is not decoded by this initial accessor.
}
