% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-sequence.R
\name{tskit_load}
\alias{tskit_load}
\title{Load a succinct tree sequence}
\usage{
tskit_load(path)
}
\arguments{
\item{path}{Existing \code{.trees} file.}
}
\value{
A \code{TreeSequence} owning the loaded native value.
}
\description{
Load a succinct tree sequence
}
