% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-sequence.R
\name{tskit_samples}
\alias{tskit_samples}
\title{Return sample node IDs}
\usage{
tskit_samples(x)
}
\arguments{
\item{x}{A \code{TreeSequence}.}
}
\value{
An integer vector of zero-based node IDs.
}
\description{
Return sample node IDs
}
