% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-sequence.R
\name{tskit_trees}
\alias{tskit_trees}
\title{Traverse tree intervals}
\usage{
tskit_trees(x)
}
\arguments{
\item{x}{A \code{TreeSequence}.}
}
\value{
A data frame containing each tree index, half-open interval, root
count, and active edge count.
}
\description{
Traverse tree intervals
}
